Fix path to the directory for temporary resources copying in save-orchestrator with Docker (#979)
* Copy resources for execution into a dedicated path on an intermediate volume, mount an intermediate volume when creating a volume for an execution
* Create an intermediate volume in tmpfs in `docker-compose.yaml`
* Change case of properties in save-orchestrator's application.properties

Actual value of `testResourcesVolumeName` should be set externally for a particular env, because both compose and swarm add a prefix to the volume name.
Test for `DockerService` are executed outside Docker, so we can't use intermediate volume mount. Same goes for local running as a plain JVM process. Using bind mount is supported for this purpose.

Follow-up after #962