Build save-frontend as a separate image (#1199)
This PR removes the dependency on save-frontend in save-backend and adds a new task `:save-frontend:buildImage`. This task uses `paketo-buildpacks/nginx` to pack all static resources along with `nginx.conf` into an image. The container exposes port 5810 and serves static resources; nginx out-of-the-box appends `E-Tag` headers to improve quality of caching.

* When deploying locally, save-frontend should be run using webpack (`:save-frontend:run*` tasks)
* Configuration has been updated both for k8s and docker-compose
* `error.html` is placed in resources of gateway and backend, because its used by spring-boot